% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_BS_PF.R
\name{adapt_BS_PF}
\alias{adapt_BS_PF}
\title{Bootstrap Particle Filter Adaptation}
\usage{
adapt_BS_PF(K0 = 10, particleFilter, theta, alpha)
}
\arguments{
\item{particleFilter}{A function which carries out Particle Filter sampling.}

\item{theta}{Epidemic parameters to be passed on to \code{particleFilter}}

\item{alpha}{Observational parameters to be passed on to \code{particleFilter}}

\item{K_0}{Initial number of particles.}
}
\value{
Returns number of particles which gives adequate variance of the log-estimate.
}
\description{
Adapts the number of particles used in a given bootstrap particle
filter to give adequate performance.
}
