% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_ISMCMC.R
\name{adapt_ISMCMC}
\alias{adapt_ISMCMC}
\title{Adapt RWM/IS MCMC}
\usage{
adapt_ISMCMC(
  init,
  U_init,
  epiModel,
  obsFrame,
  epiSample,
  I0,
  alpha,
  logPrior,
  lambda0,
  V0,
  S0,
  noIts,
  delta = 0.05
)
}
\description{
Adapt RWM/IS MCMC
}
