% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_particleMCMC.R
\name{adapt_particleMCMC}
\alias{adapt_particleMCMC}
\title{Adapt Particle MCMC}
\usage{
adapt_particleMCMC(
  init,
  epiModel,
  obsFrame,
  y,
  X0,
  alpha,
  logPrior,
  lambda0,
  V0,
  delta = 0.05,
  K,
  noIts
)
}
\description{
Adapt Particle MCMC
}
