% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseAscertainment.R
\name{caseAscObsModel}
\alias{caseAscObsModel}
\title{Case Ascertation observational model for a discrete-time meta-population SIR model.}
\usage{
caseAscObsModel(X)
}
\arguments{
\item{epidemic}{The output of the discrete-time, meta-population epidemic model}

\item{sample}{A sample generated from the testing regimes sample function}

\item{alpha}{The probability that an individual is recruited for testing
on any given day, regardless of disease status.}
}
\value{
Returns a sampling function and a density calculation function.
Sampling function for Case Ascertation regime.

Returns a cases ascertation sample from the given epidemic

\preformatted{  Returns the log-likelihood of the sample given the epidemic and
          observational parameter alpha.
}
}
\description{
Case Ascertation observational model for a discrete-time meta-population SIR model.
}
