% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_PF.R
\name{generic_PF}
\alias{generic_PF}
\title{Generic Particle Filter}
\usage{
generic_PF(y, X_0, obsFrame, epiModel, proposalFrame)
}
\arguments{
\item{y}{Observed epidemic data}

\item{X_0}{Initial epidemic state, which is assumed to be known.}

\item{obsFrame}{Generator function for observational model.}

\item{epiModel}{epidemic model}
}
\value{
Returns log-likelihood estimate.
}
\description{
Generates a particle filter which propogates particals using a given proposal scheme q.
}
