% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportanceSampler.R
\name{IS}
\alias{IS}
\title{Generate a function which calculates Monte Carlo estimate of Epidemic sample Likelihood
given a set of epidemic parameters.}
\usage{
IS(y, obsFrame, epiModel, parallel = F)
}
\arguments{
\item{obsFrame}{Function which generates the observation model given an epidemic.
A representation of how a realisation of an epidemic, described
by \code{epiModel}, is assumed to be observed.}

\item{epiModel}{Epidemic model with simulator and likelihood functions.
A representation of how the real-world epidemic is assumed
to behave.}

\item{sample}{An observation of an SIR epidemic with a metapopulation
structure. The format of this object must be the same
as the format which is returned from \code{obsModel$sample()},
where obsModel is generated by \code{obsFrame}.}
}
\description{
Generate a function which calculates Monte Carlo estimate of Epidemic sample Likelihood
given a set of epidemic parameters.
}
