% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISMCMC.R
\name{ISMCMC}
\alias{ISMCMC}
\title{Independence Sampler/RWM MCMC}
\usage{
ISMCMC(
  init,
  U_init,
  epiModel,
  obsFrame,
  epiSample,
  I0,
  alpha,
  logPrior,
  lambda,
  V,
  S,
  noIts
)
}
\arguments{
\item{init}{initial value of epidemic parameters}

\item{U_init}{initial value of random variables used to construct
epidemic trajectory.}

\item{epiModel}{epidemic model}

\item{obsFrame}{Generator function for observational model.}

\item{epiSample}{Observed epidemic data.}

\item{I0}{Initial state of the epidemic, which assumed to be known.}

\item{alpha}{Observational parameters to be passed on to the log-likelihood function
generated by \code{obsFrame(X_sim)}}

\item{logPrior}{Functions which calculate log-density of the assumed priors of the
epidemic parameters.}

\item{lambda}{Random Walk Metropolis (RWM) proposal scale parameter.}

\item{V}{A square matrix containing the covariance values for RWM proposal. Shape
of matrix should match the length of the init parameter.}

\item{S}{Number of non-centered random variables to be 'refreshed' in Independence
Sampler proposal step.}

\item{noIts}{Number of iterations of MCMC sampler scheme to carry out.}
}
\value{
A matrix of MCMC samples and proposal acceptance rate of the sample.
}
\description{
Generates dependent samples from target distribution
using a Independence Sampler/RWM MCMC proposal scheme.
}
