% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{A function which adjusts average likelihood calculation to
avoid numerical error which leads to false zeroes
THIS METHOD DOES NOT WORK IN SOME SITUATIONS
May make a adjustments such that exp(a_i + b) -> Inf}
\usage{
log_sum_exp(a)
}
\arguments{
\item{a}{vector of log-likelihood values.}
}
\value{
The log(likelihood-estimate)
}
\description{
A function which adjusts average likelihood calculation to
avoid numerical error which leads to false zeroes
THIS METHOD DOES NOT WORK IN SOME SITUATIONS
May make a adjustments such that exp(a_i + b) -> Inf
}
