% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSIR.R
\name{metaSIR}
\alias{metaSIR}
\title{Discrete-time SIR epidemic model}
\usage{
metaSIR(N_M, endTime)
}
\arguments{
\item{N_M}{The size of each metapopulation. The length of this
vector will determine how many metapopulations there
are.}

\item{endTime}{At what time does simulation of the epidemic
stop. By default, endTime is infinite meaning
the epidemic is simulated until it dies out.}

\item{epidemic}{An epidemic realisation which is in the format
returned by \code{sim}.}

\item{theta}{Epidemic parameters. Three parameters need to be
given corresponding to beta_G, beta_L and
gamma in that order (see \code{dailyProg}).}
}
\value{
A \code{sim} function to simulate from the generated model,
given a set of parameters. A \code{llh} function to calculate
the log-density for an epidemic, given a set of parameters.
Calculate the log-likelihood of an SIR epidemic which
is assumed to take place in a metapopulation structure
occurs, given a parameter set.

\preformatted{    Returns log-likelihood value corresponding to
            the given epidemic and parameter set.
}
}
\description{
Returns functions to simulate or calculate the log-density of a
discrete-time SIR Process with a metapopulation structure,
given a parameter set.
}
