% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSIR_NC.R
\name{metaSIR_NC}
\alias{metaSIR_NC}
\title{Generate a non-centered discrete-time epidemic model}
\usage{
metaSIR_NC(N_M, startTime, endTime, PRINT = F)
}
\arguments{
\item{N_M}{Size of each meta-population.}

\item{endTime}{The time which the epidemic should be simulated for.}
}
\value{
Returns three functions. A \code{sim} function to simulate from the generated model,
given a set of parameters. A \code{dailyProg} function, whose behaviour is identical
to \code{sim} but simulates for one time unit only. Finally, a \code{llh} function to calculate
he log-density for an epidemic, given a set of parameters.
}
\description{
Generates a discrete-time epidemic model which constructs epidemics
using a stream of random variables which are independent to the
epidemic parameters and state.
}
