% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particleMCMC.R
\name{particleMCMC}
\alias{particleMCMC}
\title{Particle MCMC}
\usage{
particleMCMC(
  init,
  epiModel,
  obsFrame,
  y,
  X0,
  alpha,
  logPrior,
  lambda,
  V,
  K,
  noIts
)
}
\arguments{
\item{init}{initial value of epidemic parameters}

\item{epiModel}{epidemic model}

\item{obsFrame}{Generator function for observational model.}

\item{y}{Observed epidemic data.}

\item{X0}{Initial state of the epidemic, which assumed to be known.}

\item{alpha}{Observational parameters to be passed on to the log-likelihood function
generated by \code{obsFrame(X_sim)}}

\item{logPrior}{Functions which calculate log-density of the assumed priors of the
epidemic parameters.}

\item{lambda}{Random Walk Metropolis (RWM) proposal scale parameter}

\item{V}{A square matrix containing the covariance values for RWM proposal. Shape
of matrix should match the length of the init parameter.}

\item{K}{Number of particles to be used in particle filter. This can be optimised using
\code{adapt_BS_PF()}.}

\item{noIts}{Number of iterations of MCMC sampler scheme to carry out.}
}
\value{
A matrix of MCMC samples and proposal acceptance rate of the sample.
}
\description{
Generates dependent samples from target distribution
using a Particle MCMC proposal scheme.
}
