% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsMCMC.R
\name{PsMCMC}
\alias{PsMCMC}
\title{Pseudo-Marginal MCMC Sampler for epidemic models}
\usage{
PsMCMC(
  init,
  epiModel,
  obsFrame,
  epiSample,
  I0,
  alpha,
  logPrior,
  lambda,
  V,
  N,
  noIts
)
}
\description{
Pseudo-Marginal MCMC Sampler for epidemic models
}
