% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascCase.R
\name{r/dcaseAscObsModel}
\alias{r/dcaseAscObsModel}
\alias{rAscCases}
\title{Case Ascertainment observational model for a discrete-time meta-population SIR model.}
\usage{
rAscCases(alpha, X)
}
\arguments{
\item{X}{Underlying epidemic trajectory}
}
\value{
Returns a sampling function and a log-density calculation function.
}
\description{
Generates a case ascertainment observation model. There is assumed
to be a fixed probability alpha of observing an infection, independent
of time and individual.
}
