% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_to_binom.R
\name{unif_to_binom}
\alias{unif_to_binom}
\title{Transform Uniform Random Variable into a Binomial Random Variable}
\usage{
unif_to_binom(U, trials, prob)
}
\description{
Transform Uniform Random Variable into a Binomial Random Variable
}
