% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_bootstrap_PF.R
\name{VBS_PF}
\alias{VBS_PF}
\title{Bootstrap Particle Filter using vectorised simulator}
\usage{
VBS_PF(y, X_0, epiModel, obs = "AC")
}
\arguments{
\item{y}{Observed epidemic data}

\item{X_0}{Initial epidemic state, which is assumed to be known.}

\item{epiModel}{epidemic model}

\item{obsFrame}{Generator function for observational model.}
}
\value{
Returns log-likelihood estimate.
}
\description{
Generates a bootstrap particle filter.
}
\details{
Bootstrap Particle Filter using vectorised simulator
}
