% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_particleMCMC.R
\name{adapt_particleMCMC}
\alias{adapt_particleMCMC}
\title{Particle MCMC Adaptation}
\usage{
adapt_particleMCMC(
  init,
  epiModel,
  obsFrame,
  y,
  X0,
  alpha,
  logPrior,
  lambda0,
  V0,
  delta = 0.05,
  K,
  noIts
)
}
\arguments{
\item{init}{initial value of epidemic parameters}

\item{epiModel}{epidemic model}

\item{obsFrame}{Generator function for observational model.}

\item{y}{Observed epidemic data.}

\item{X0}{Initial state of the epidemic, which assumed to be known.}

\item{alpha}{Observational parameters to be passed on to the log-likelihood function
generated by \code{obsFrame(X_sim)}}

\item{logPrior}{Functions which calculate log-density of the assumed priors of the
epidemic parameters.}

\item{lambda0}{Initial value of the Random Walk Metropolis (RWM) proposal scale parameter}

\item{V0}{A square matrix containing the initial covariance values for RWM proposal. Shape
of matrix should match the length of the init parameter.}

\item{delta}{Probability that the initial proposal parameters are used as opposed to using
the current adapted proposal parameters.}

\item{K}{Number of particles to be used in particle filter. This can be optimised using
\code{adapt_BS_PF()}.}

\item{noIts}{Number of iterations of MCMC sampler scheme to carry out.}
}
\value{
A list of the adapted proposal parameters.
}
\description{
Adapts proposal parameters of Particle MCMC scheme
to give near-optimal performance.
}
