% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportanceSampler.R
\name{IS}
\alias{IS}
\title{Importance Sampler}
\usage{
IS(y, obsFrame, epiModel, parallel = F)
}
\arguments{
\item{obsFrame}{Generator function for assumed observational model.}

\item{epiModel}{epidemic model.}

\item{sample}{Observed epidemic data.}
}
\value{
An Importance Sampler function.
}
\description{
Generate a function which calculates Monte Carlo estimate of Epidemic sample Likelihood
given a set of epidemic parameters.
}
