% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISMCMC.R
\name{ISMCMC}
\alias{ISMCMC}
\title{Importance Sampling MCMC for non-centered epidemic models}
\usage{
ISMCMC(
  init,
  U_init,
  epiModel,
  obsFrame,
  epiSample,
  I0,
  alpha,
  logPrior,
  lambda,
  V,
  S,
  noIts
)
}
\description{
Importance Sampling MCMC for non-centered epidemic models
}
