% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particleMCMC.R
\name{particleMCMC}
\alias{particleMCMC}
\title{Particle MCMC}
\usage{
particleMCMC(
  init,
  epiModel,
  obsFrame,
  y,
  X0,
  alpha,
  logPrior,
  lambda,
  V,
  K,
  noIts
)
}
\description{
Particle MCMC
}
