% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testingObsModel}
\alias{testingObsModel}
\title{Case Ascertainment observational model for a discrete-time meta-population SIR model.}
\usage{
testingObsModel(epidemic)
}
\arguments{
\item{epidemic}{Underlying epidemic trajectory}
}
\value{
Returns a sampling function and a log-density calculation function.
}
\description{
Generates a testing observation model. There is assumed
to be a fixed probability \alpha of an individual being recruited for testing, independent
of time and individual. There is then an associated sensitivity and specificity for the test.
}
